requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
brotli==1.2.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
ddgs==9.9.1
distro==1.9.0
et_xmlfile==2.0.0
fake-useragent==2.2.0
greenlet==3.2.4
griffe==1.15.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
markdown-it-py==4.0.0
mcp==1.22.0
mdurl==0.1.2
numpy==2.3.5
openai==2.8.1
openai-agents==0.6.1
openpyxl==3.1.5
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
playwright==1.56.0
primp==0.15.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pypdfium2==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
referencing==0.37.0
requests==2.32.5
rich==14.2.0
rpds-py==0.29.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sse-starlette==3.0.3
starlette==0.50.0
tqdm==4.67.1
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
